the_game/requirements.txt
Flask==2.1.3
Werkzeug==2.3.7
Flask-CORS==3.0.10
requests==2.27.1
mysql-connector-python==8.0.28
python-dotenv==0.19.2
geopy==2.2.0